+++ R/taxlist-class.R
@@ -30,7 +30,9 @@ setClass("taxlist",
                 ),
                 hierarchy=character(),
                 taxonViews=data.frame(
-                        ViewID=integer()
+                        ViewID=integer(),
+						Secundum=character(),
+						stringsAsFactors=FALSE
                 ),
 				taxonTraits=data.frame(
                         TaxonConceptID=integer()
@@ -60,8 +62,8 @@ setClass("taxlist",
                             object@taxonRelations$TaxonConceptID))
                 return("Some concepts are missing in slot 'taxonRelations'")
 			# taxonView
-            if(!"ViewID" %in% colnames(object@taxonViews))
-                return("'ViewID' is a mandatory column in slot 'taxonViews'")
+			if(!all(c("ViewID","Secundum") %in% colnames(object@taxonViews)))
+	            return("'ViewID' and 'Secundum' are a mandatory columns in slot 'taxonViews'")
             if(any(duplicated(object@taxonViews$ViewID)))
                 return("Duplicated views are not allowed in slot 'taxonViews'")
             if(nrow(object@taxonViews) > 0 &