+++ NEWS.md
@@ -5,6 +5,7 @@ Current
 
 * New function `load_last` to load last backup in an R-session.
 * File **inst/ChangeLog** replaced by **NEWS.md**.
+* New column `Secundum` as mandatory field in slot `taxonViews`.
 
 ### Improvements